Revert "Restructure stuff"
